+++ program/src/state.rs
@@ -1,3 +1,3 @@
 //! Program state types.
 
-pub use solana_config_interface::ConfigKeys;
+pub use solana_config_interface::state::ConfigKeys;
